+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+v0.4.8
+- Migrate Guild handler to database
+- Create `network` module
+- Back to twitter-scraper
+- Update multitor
+
 v0.4.7
 - Fix `Update` command out of array
 - Change 404.jpg Thumbnail to null